API helpers - enrichment and replication
